Refactoring:  - Add Global action to go to Login destination  - Add base fragment with the logic to the menu
