+++ app/src/main/java/br/com/alura/aluraesporte/ui/fragment/BaseFragment.kt
@@ -1,0 +1,72 @@
+package br.com.alura.aluraesporte.ui.fragment
+
+import android.os.Bundle
+import android.view.Menu
+import android.view.MenuInflater
+import android.view.MenuItem
+import android.view.View
+import androidx.core.view.MenuHost
+import androidx.core.view.MenuProvider
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.Lifecycle
+import androidx.navigation.fragment.findNavController
+import br.com.alura.aluraesporte.NavGraphDirections
+import br.com.alura.aluraesporte.R
+import br.com.alura.aluraesporte.ui.viewmodel.LoginViewModel
+import org.koin.android.viewmodel.ext.android.viewModel
+
+abstract class BaseFragment:  Fragment() {
+
+    protected val loginViewModel: LoginViewModel by viewModel()
+    protected val navController by lazy {
+        findNavController()
+    }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        isLogged()
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        setUpMenu()
+    }
+
+    private fun setUpMenu() {
+        val menuHost: MenuHost = requireActivity()
+        menuHost.addMenuProvider(object : MenuProvider {
+            override fun onCreateMenu(menu: Menu, menuInflater: MenuInflater) {
+                menuInflater.inflate(R.menu.menu_main, menu)
+            }
+
+            override fun onMenuItemSelected(menuItem: MenuItem): Boolean {
+                return when (menuItem.itemId) {
+                    R.id.menu_logout -> {
+                        loginViewModel.logOut()
+                        goToLogin()
+                        true
+                    }
+
+                    else -> false
+                }
+            }
+
+        }, viewLifecycleOwner, Lifecycle.State.RESUMED)
+    }
+
+
+    private fun isLogged() {
+        if (!loginViewModel.isLogged()) {
+            goToLogin()
+        }
+    }
+
+    protected fun goToLogin() {
+        val directions =
+            NavGraphDirections.actionGlobalLogin()
+        navController.navigate(directions)
+    }
+
+
+
+}
