+++ app/src/main/res/menu/menu_main.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
-    <item android:id="@+id/menu_products_list_logout"
+    <item android:id="@+id/menu_logout"
         android:title="@string/log_out"
         android:icon="@drawable/ic_action_logout"
         app:showAsAction="always"
