+++ app/src/main/res/navigation/nav_graph.xml
@@ -79,6 +79,15 @@
             app:popUpTo="@id/login"
             app:popUpToInclusive="true" />
     </fragment>
+    <action
+        android:id="@+id/action_global_login"
+        app:destination="@id/login"
+        app:enterAnim="@anim/nav_default_enter_anim"
+        app:exitAnim="@anim/nav_default_exit_anim"
+        app:popEnterAnim="@anim/nav_default_pop_enter_anim"
+        app:popExitAnim="@anim/nav_default_pop_exit_anim"
+        app:popUpTo="@id/productsList"
+        app:popUpToInclusive="true"/>
 
 
 </navigation>
